added adding activity and recording of grades
To Fix: Attendance

To Do: Admin and Student

+++ src/main/java/com/assookkaa/ClassRecord/Dto/Request/GradingComposition/GradingCompositionDtoRequest.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class GradingCompositionDtoRequest {
-    private Float percentage;
+    private Double percentage;
     private Integer categoryId;
     private Integer teachingLoadDetailId;
 }
